bugfix tx_scale
Former-commit-id: 75a7b3d271008500203c3c737eb95b8f71d5730e

+++ CC/Sounder/include/config.h
@@ -48,7 +48,7 @@ public:
     int dlSymsPerFrame;
     int fftSize;
     int cpSize;
-    double tx_scale;
+    float tx_scale;
     std::string pilot_seq;
     std::string beacon_seq;
